Relocate sub components into own folder
