+++ app/components/UI/SimulationDetails/BalanceChangeList/BalanceChangeList.styles.ts
