+++ app/components/UI/SimulationDetails/BalanceChangeList/BalanceChangeList.test.tsx
@@ -3,12 +3,12 @@ import { render } from '@testing-library/react-native';
 import { BigNumber } from 'bignumber.js';
 
 import BalanceChangeList from './BalanceChangeList';
-import { BalanceChange } from './types';
-import { sortBalanceChanges } from './sortBalanceChanges';
-import BalanceChangeRow from './BalanceChangeRow';
+import { BalanceChange } from '../types';
+import { sortBalanceChanges } from '../sortBalanceChanges';
+import BalanceChangeRow from '../BalanceChangeRow/BalanceChangeRow';
 
-jest.mock('./sortBalanceChanges');
-jest.mock('./BalanceChangeRow', () => 'BalanceChangeRow');
+jest.mock('../sortBalanceChanges');
+jest.mock('../BalanceChangeRow/BalanceChangeRow', () => 'BalanceChangeRow');
 
 const balanceChangesMock = [
   {
