+++ app/components/UI/SimulationDetails/BalanceChangeRow/BalanceChangeRow.styles.ts
