+++ app/components/UI/SimulationDetails/BalanceChangeRow/BalanceChangeRow.tsx
@@ -3,13 +3,13 @@ import React from 'react';
 import { View, ViewProps } from 'react-native';
 import Text, {
   TextVariant,
-} from '../../../component-library/components/Texts/Text';
-import { useStyles } from '../../hooks/useStyles';
+} from '../../../../component-library/components/Texts/Text';
+import { useStyles } from '../../../hooks/useStyles';
 
 import styleSheet from './BalanceChangeRow.styles';
-import { BalanceChange } from './types';
-import AmountPill from './AmountPill';
-import AssetPill from './AssetPill';
+import { BalanceChange } from '../types';
+import AmountPill from '../AmountPill/AmountPill';
+import AssetPill from '../AssetPill/AssetPill';
 
 interface BalanceChangeRowProperties extends ViewProps {
   label?: string;
